Settings: recover last location state during on/off switch
based on the behaviour change on QS here:

https://gerrit.slimroms.net/#/c/1421

we go now same way in settings. Tbh not sure what the goal is to
force the user after switching off to enable always high accuracy
location. Whatever we use here same attempt like in QS tile....if the
user switches on again it recovers the last enabled state.

Change-Id: I887e2a94af379770dac7dbedb5d3d8002a554163